requirements.txt
flask
PyMySQL
mysql-connector-python==8.0.11